pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "decision-agent"
version = "2.0.0"
description = "Complete Decision Agent with intelligent task routing"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Decision Agent Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "aiohttp>=3.9.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "openai>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/decision-agent"
"Bug Tracker" = "https://github.com/yourusername/decision-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["decision_agent*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true